+++ src/main/java/dev/book/global/config/security/dto/oauth2/OAuth2Attributes.java
@@ -1,4 +1,4 @@
-package dev.book.global.config.security.dto;
+package dev.book.global.config.security.dto.oauth2;
 
 import lombok.Builder;
